sanitization, validation and escaping (#1087)
* fix: removed set time limit

* refactor: removed redundant files and unsanitized data

* refactor: added sanitizations to global vars

* refactor: added sanitizations to global vars

* fix: deregistered plugin_apis hook

* Revert "fix: deregistered plugin_apis hook"

This reverts commit ad6d45b6ebb153ff981819821b98e7e15f268cf2.

* fix: removed redundant files and screenshot

* fix: finalized escaping

+++ includes/api/class-urlslab-api-faq.php
@@ -236,6 +236,13 @@ class Urlslab_Api_Faq extends Urlslab_Api_Table {
 	}
 
 	public function delete_all_items( WP_REST_Request $request ) {
+		//# Sanitization
+		$sanitized_req = $request->sanitize_params();
+		if ( is_wp_error( $sanitized_req ) ) {
+			return $sanitized_req;
+		}
+		//# Sanitization
+
 		global $wpdb;
 
 		if ( false === $wpdb->query( $wpdb->prepare( 'TRUNCATE ' . sanitize_key( $this->get_row_object()->get_table_name() ) ) ) ) { // phpcs:ignore
